requirements.txt
streamlit
snowflake-connector-python
snowflake-snowpark-python
cryptography
pandas
altair
